backend/requirements.txt
fastapi==0.86.0
httptools==0.5.0
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
pydantic==1.10.2
python-multipart==0.0.5
starlette==0.20.4
urllib3==1.26.12
uvicorn==0.19.0
stability-sdk